Cleanup plugin.py, move harvester functions into its own package
